Rename l10n_es_tgss_base to l10n_es_tgss.
