+++ l10n_es_tgss/README.rst
@@ -64,7 +64,7 @@ Bugs are tracked on `GitHub Issues <https://github.com/OCA/
 l10n-spain/issues>`_. In case of trouble, please check there if your issue has
 already been reported. If you spotted it first, help us smashing it by
 providing a detailed and welcomed feedback `here <https://github.com/OCA/
-l10n-spain/issues/new?body=module:%20 l10n_es_tgss_base%0Aversion:%20
+l10n-spain/issues/new?body=module:%20 l10n_es_tgss%0Aversion:%20
 8.0.1.0.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
