+++ l10n_es_tgss/exceptions.py
