+++ l10n_es_tgss/models/common.py
@@ -2,4 +2,4 @@
 # © 2015 Grupo ESOC Ingeniería de servicios, S.L.
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-_NS = "l10n_es_tgss_base"
+_NS = "l10n_es_tgss"
