+++ l10n_es_tgss/models/professional_category.py
