+++ l10n_es_tgss/models/special_situations.py
