+++ l10n_es_tgss/security/ir.model.access.csv
@@ -1,0 +1,7 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_contribution_group_manager,Contribution Groups Manager,model_l10n_es_tgss_contribution_group,base.group_system,1,1,1,1
+access_contribution_group_user,Contribution Groups User,model_l10n_es_tgss_contribution_group,base.group_partner_manager,1,0,0,0
+access_professional_category_manager,Professional category manager,model_l10n_es_tgss_professional_category,base.group_system,1,1,1,1
+access_professional_category_user,Professional category user,model_l10n_es_tgss_professional_category,base.group_partner_manager,1,0,0,0
+access_study_level_manager,Study Level Manager,model_l10n_es_tgss_study_level,base.group_system,1,1,1,1
+access_study_level_user,Study Level User,model_l10n_es_tgss_study_level,base.group_partner_manager,1,0,0,0
