+++ l10n_es_tgss/tests/disability_percentage.py
