+++ l10n_es_tgss/views/professional_category.xml
@@ -5,7 +5,7 @@
 <!-- Views -->
 <record id="professional_category_list" model="ir.ui.view">
     <field name="name">Professional Category List</field>
-    <field name="model">l10n_es_tgss_base.professional_category</field>
+    <field name="model">l10n_es_tgss.professional_category</field>
     <field name="priority" eval="1"/>
     <field name="arch" type="xml">
         <tree string="Professional Category List" editable="bottom">
@@ -17,7 +17,7 @@
 
 <record id="professional_category_search" model="ir.ui.view">
     <field name="name">Professional Category Search</field>
-    <field name="model">l10n_es_tgss_base.professional_category</field>
+    <field name="model">l10n_es_tgss.professional_category</field>
     <field name="priority" eval="1"/>
     <field name="arch" type="xml">
         <search>
@@ -31,7 +31,7 @@
 <record id="professional_category_list_action" model="ir.actions.act_window">
     <field name="name">Professional categories</field>
     <field name="type">ir.actions.act_window</field>
-    <field name="res_model">l10n_es_tgss_base.professional_category</field>
+    <field name="res_model">l10n_es_tgss.professional_category</field>
     <field name="view_type">form</field>
     <field name="view_mode">tree</field>
     <field name="view_id" ref="professional_category_list"/>
