+++ l10n_es_tgss_partner/tests/test_disability_percentage.py
@@ -3,7 +3,7 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 from .base import Contact
-from openerp.addons.l10n_es_tgss_base.tests import disability_percentage
+from openerp.addons.l10n_es_tgss.tests import disability_percentage
 
 
 class DisabilityPercentageCase(Contact,
